[refactor] [#1] 너무 큰 컴포넌트 쪼개기
- 너무 크다고 생각되는 컴포넌트를 하위 컴포넌트 몇개로 쪼갬
